requirements.txt
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-admin-thumbnails==0.2.7
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-grappelli==3.0.6
django-health-check==3.17.0
djangorestframework==3.14.0
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2==2.9.2
pycodestyle==2.10.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
regex==2023.5.5
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
